+++ docs/CHANGELOG.rst
@@ -20,6 +20,8 @@ Changed
 - Add parameter ``--use-original-qualities`` to ``bqsr`` process
 - Add ``fn_ignore_dirs`` options to MultiQC configuration file in
   ``resolwebio/common:3.0.1`` Docker image
+- Add parameter ``--exclude-filtered`` to ``gatk-select-variants``
+  process
 
 Fixed
 -----